moved object to dedicated edit_conf module within package
