+++ heteropolymer_simulations/__init__.py
@@ -6,6 +6,7 @@ A package for running terphenyl oligomery systems
 import heteropolymer_simulations.utils
 import heteropolymer_simulations.observables
 import heteropolymer_simulations.scripts
+import heteropolymer_simulations.edit_conf
 
 # Handle versioneer
 from ._version import get_versions
